Continue to move AlphaScraper to individual table classes and a separate API class (WIP)
