+++ esgtools/etl/etl_alpha_api.py
@@ -1,0 +1,52 @@
+
+import os
+import requests
+import time
+import pandas as pd
+from datetime import datetime
+
+
+URL_BASE = 'https://www.alphavantage.co/query?function='
+
+
+class AlphaScraper():
+
+    def __init__(self, wait=True, max_api_requests_per_min=75):
+        self.api_key = os.environ.get('ALPHAVANTAGE_API_KEY')
+        self.wait = wait
+        self.max_api_requests_per_min = max_api_requests_per_min
+        self.remaining_api_requests = max_api_requests_per_min
+        self.last_increase_time = datetime.today()
+
+    def wait_to_hit_api(self):
+
+        # Remaining api requests
+        if self.remaining_api_requests > 0:
+            self.remaining_api_requests -= 1
+            print(f"Remaining api requests: {self.remaining_api_requests}")
+            return
+        
+        # Reached maximum
+        delay = (datetime.today() - self.last_increase_time).total_seconds()
+        if delay < 60:
+            print(f"Maximum limit reached. Sleeping for {60 - delay} seconds")
+            time.sleep(60 - delay)
+        
+        print("API limit refreshed after 1 minute")
+        self.remaining_api_requests = self.max_api_requests_per_min - 1
+        self.last_increase_time = datetime.today()
+
+
+    def hit_api(self, url, **kwargs):
+
+        kwargs["URL_BASE"] = URL_BASE
+        kwargs["api_key"] = self.api_key
+        url = url.format(**kwargs)
+        
+        if self.wait:
+            self.wait_to_hit_api()
+
+        with requests.Session() as s:
+            download = s.get(url)
+
+        return download
